move ddm test to their own file
